// src/components/types/Game.ts
interface Game {
    Title: string;
    Year: string;
    imdbID: string;
    Type: string;
    Poster: string;
}

interface GameDetail {
    Title: string;
    Year: string;
    Rated: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    Writer: string;
    Actors: string;
    Plot: string;
    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Ratings: { Source: string; Value: string }[];
    Metascore: string;
    imdbRating: string;
    imdbVotes: string;
    imdbID: string;
    DVD: string;
    BoxOffice: string;
    Production: string;
    Website: string;
}

export type { Game, GameDetail };